Make loading data
URL segments, layouts, and data are more often than not coupled (tripled?) together.

Because of this natural coupling, React Router has data conventions to get data into your route components easily.

There are two APIs we'll be using to load data, loader and useLoaderData. First we'll create and export a loader function in the root module, then we'll hook it up to the route. Finally, we'll access and render the data.

Order:
- Export a loader from root.jsx
- Configure the loader on the route
- Access and render the data

React Router will now automatically keep that data in sync with your UI. We don't have any data yet, so you're probably getting a blank list