Add updating data by clicking "Edit"
Just like creating data, you update data with <Form>. Let's make a new route at contacts/:contactId/edit. Again, we'll start with the component and then wire it up to the route config.

Order:
- Create the edit component
- Add the edit page UI
- Add the new edit route

We want it to be rendered in the root route's outlet, so we made it a sibling to the existing child route.

(You might note we reused the contactLoader for this route. This is only because we're being lazy in the tutorial. There is no reason to attempt to share loaders among routes, they usually have their own.)

Alright, clicking the "Edit" button gives us this new UI